Add files via upload
ExtraaLearn is an initial stage startup that offers programs on cutting-edge technologies to students and professionals to help them upskill/reskill. With a large number of leads being generated regularly, one of the issues faced by ExtraaLearn is to identify which of the leads are more likely to convert so that they can allocate resources accordingly. You, as a data scientist at ExtraaLearn, have been provided the leads data to:

    Analyze and build an ML model to help identify which leads are more likely to convert to paid customers,
    Find the factors driving the lead conversion process
    Create a profile of the leads which are likely to convert